22092026 — Full frontend redesign, backend streamline, perf + security
Theme:
- PDP redesign: swipe gallery, offers strip, virtual tour, pros/cons cards
- Homepage: developer partners marquee, stats counter, hero chips centered
- Homepage perf: transient caching for all queries (6hr TTL), cache invalidation on save_post
- New pages: About (founders, vision/mission), Careers, Contact (with AJAX forms)
- Blog system: single-tp_guide, archive, blog page, sidebar CTA with AJAX lead capture
- Search/taxonomy archives: redesigned cards, filter bar
- Footer: redesigned with social links, RERA info
- Header: post property modal, mobile menu
- Banner images: auto-crop to 800x530 on upload
- Similar projects: overlay carousel arrows

Plugin:
- Admin: streamlined metaboxes (removed Construction, Pricing, Scoring tabs)
- Landing page CPT with root-level URL resolver
- Guide CPT for blog posts
- Location taxonomy: homepage checkbox meta
- Settings: brand settings admin page
- Lead API: public endpoint for forms

Security:
- .htaccess: GZIP compression, security headers, hotlink protection, bot blocking
- Static asset caching: 1yr immutable headers

+++ themes/tenprojects/template-parts/project/commercial-specs.php
@@ -24,7 +24,7 @@ $fire_noc          = tp_get_meta( $post_id, 'fire_noc' );
 $lease_term        = tp_get_meta( $post_id, 'lease_term' );
 $lock_in_period    = tp_get_meta( $post_id, 'lock_in_period' );
 $escalation_clause = tp_get_meta( $post_id, 'escalation_clause' );
-$price_min         = intval( tp_get_meta( $post_id, 'price_display_min' ) );
+$price_min         = floatval( tp_get_meta( $post_id, 'price_display_min' ) );
 
 $has_data = $commercial_type || $commercial_carpet || $price_per_sqft || $price_min;
 if ( ! $has_data ) {
